requirements.txt
Django==3.2.6
djangorestframework==3.12.4
django-debug-toolbar==3.2.2